tmpfiles: re-add /var/run symlink rule
This is a quick workaround for a regression caused by 54728a4a which
removed the conflicting rule. This changed behavior slightly because
bootengine calls tmpfiles with baselayout.conf but not systemd's
var.conf which has the other rule and something (possibly dbus) was
creating /var/run as a directory before the standard tmpfiles service
created it as a symlink.

Ideally dbus should be fixed to use /run instead of /var/run so we can
reduce the amount of magic bootengine has to do but for now just re-add
the rule, but matching the other so that they don't conflict.