tmpfiles: move copy rule to new config
Copy rules are always use the given absolute path, regardless of --root.
This means that when baselayout-etc.conf is executed in the initrd it
fails because /usr/share/baselayout is not included in the initrd. Move
the copy rule to a new config that does not get executed so early.