pam: order pam_sss after pam_unix in auth section
The sssd documentation always orders pam_sss after pam_unix. We already
do so in all the other sections. Given the use_first_pass option is
given (re-use auth token from previous plugin) it is clearly intended to
not be the first. In the inverted order it seems like using sssd for
password based auth will not work. Clearly we need to get sssd into our
automated testing in some way or another.