baselayout: Allow pam_sss to reject users
sssd has support for whitelisting or blacklisting users and groups. To do
so, pam_sss must be able to block login, which means it has to be listed as
required in the account section of sys-auth. The ignore_unknown_user and
ignore_authinfo_unavail mean that it will not block logins if the user is
unknown by sssd or if pam_sss is unable to communicate with sssd.