move user/group definitions into sysusers
This commit removes almost all definitions from baselayout/passwd and
baselayout/group, and adds these definitions into a new file,
sysusers.d/coreos.conf. This new file is then used by sysusers to create
these users. This happens a few times, once during the bootengine build
to create users and groups in the initramfs, once during the initramfs
on the first boot to create users and groups in the /sysroot partition,
and every time on boot to apply any potential changes to the sysusers
definitions.