coreos-profile: Show update status using conf file
This coordinator conf file is being introduced in locksmith and the
Container Linux Update Operator.

This allows the profile to correctly display information about the
update strategy regardless whether it's being managed by locksmithd or
cluo.

See https://github.com/coreos/container-linux-update-operator/issues/88

This retains the same message for the case of a coordinator not running.

It adds additional messages for other statuses, such as "disabled" and
"reboot planned".
These messages are highlighted in yellow.