sysctl: Revert new protections from systemd 241
With systemd 241, new Linux 4.19 protections are enabled so that
e.g. root can't open files with O_CREAT in /tmp if they are owned
by another user.  This security improvement is also a backwards
incompatible change, so we're opting to preserve existing behavior
at this point in the Container Linux life cycle.